removed separate dmg target
